Self HEaling Infrastructure/requirements.txt
# SelfHealing Sentinel - Python Dependencies

# Core dependencies
python-dateutil>=2.8.2
pytz>=2021.3
PyYAML>=6.0
requests>=2.27.1
typing-extensions>=4.1.1

# Web framework
flask>=2.0.3
gunicorn>=20.1.0
Werkzeug>=2.0.3

# Monitoring and metrics
prometheus-client>=0.14.1
prometheus-flask-exporter>=0.20.2

# Machine learning and data processing
numpy>=1.22.3
pandas>=1.4.1
scikit-learn>=1.0.2
scipy>=1.8.0
statsmodels>=0.13.2

# Kubernetes integration
kubernetes>=23.3.0

# Visualization (for development/debugging)
matplotlib>=3.5.1
seaborn>=0.11.2

# Testing
pytest>=7.0.1
pytest-cov>=3.0.0

# Development tools
black>=22.1.0
flake8>=4.0.1
isort>=5.10.1
mypy>=0.941

# Documentation
Sphinx>=4.4.0
sphinx-rtd-theme>=1.0.0

# Async support
asyncio>=3.4.3
aiohttp>=3.8.1

# CLI tools
click>=8.0.4
tqdm>=4.63.0

# Security
cryptography>=36.0.1

# Time series analysis
prophet>=1.0.1

# Alerting and notifications
slack-sdk>=3.15.2